content js 를 단지 main 를 부르기 위해 수정
모든 로직은 main js 에 있음
